+++ CHANGELOG.md
@@ -6,15 +6,24 @@
 
 #### Features
 
-- Updated to govuk-frontend@3.11.0
-  - See https://github.com/alphagov/govuk-frontend/releases/tag/v3.11.0 for full release notes
-  - Cookie banner component added
-
 #### Breaking changes
 
 ---
 
 ## Releases
+
+### v5.1.0
+
+#### Fixes
+
+- Fix overeager instantiation of govuk JS (https://github.com/surevine/govuk-react-jsx/pull/110)
+  - Fixes https://github.com/surevine/govuk-react-jsx/issues/99
+
+#### Features
+
+- Updated to govuk-frontend@3.11.0
+  - See https://github.com/alphagov/govuk-frontend/releases/tag/v3.11.0 for full release notes
+  - Cookie banner component added
 
 ### v5.0.0